chore: rename internal/observable to internal/instance
