+++ packages/monix-reactive/src/internal/builders/empty.js.flow
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-import { ObservableInstance } from "../observable"
+import { ObservableInstance } from "../instance"
 
 /**
  * Empty observable object: bottom type, type parameter not used (phantom type)
