+++ packages/monix-reactive/src/internal/builders/eval.js.flow
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-import { ObservableInstance } from "../observable"
+import { ObservableInstance } from "../instance"
 
 declare export class EvalAlwaysObservable<+A> extends ObservableInstance<A> {
   constructor(_fn: () => A): EvalAlwaysObservable;
