+++ packages/monix-reactive/src/internal/builders/loop.ts
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-import { ObservableInstance } from "../observable"
+import { ObservableInstance } from "../instance"
 import { Continue, Stop, AsyncAck } from "../../ack"
 import { Subscriber } from "../../observer"
 import { Cancelable, Scheduler, BoolCancelable, IBoolCancelable } from "funfix"
