+++ packages/monix-reactive/src/internal/builders/pure.js.flow
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-import { ObservableInstance } from "../observable"
+import { ObservableInstance } from "../instance"
 
 declare export class PureObservable<+A> extends ObservableInstance<A> {
   constructor(_value: A): PureObservable;
